requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic[dotenv]
langchain
rank-bm25
chromadb
faiss-cpu
sentence-transformers
transformers
torch
scikit-learn
nltk
tiktoken
python-dotenv

tenacity>=8.2.3

python-dotenv>=1.0.1
